Holdings finished
