+++ api/services/holding/store_createholding.go
@@ -3,6 +3,7 @@ package holding
 import (
 	"context"
 	"errors"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
@@ -18,10 +19,13 @@ func (s *PostgresHoldingStore) CreateHolding(ctx context.Context, h *types.Holdi
 
 	row := s.db.QueryRow(
 		c,
-		`INSERT INTO holdings
-		(name, ticker, asset_category, expense_ratio_pct, maturation_date, interest_rate_pct, is_deprecated, user_id)
-		VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
-		RETURNING *`,
+		fmt.Sprintf(`
+			insert into holdings
+				(name, ticker, asset_category, expense_ratio_pct, maturation_date, interest_rate_pct, is_deprecated, user_id)
+				values ($1, $2, $3, $4, $5, $6, $7, $8)
+			returning
+				%s
+		`, holdingsColumns),
 		h.Name, h.Ticker, h.Asset_category, h.Expense_ratio_pct, h.Maturation_date, h.Interest_rate_pct, h.Is_deprecated, h.User_id,
 	)
