+++ api/services/holding/store_getholdings.go
@@ -10,7 +10,6 @@ import (
 	"github.com/MicroFish91/portfolio-instruments-api/api/querybuilder"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
 	"github.com/MicroFish91/portfolio-instruments-api/api/utils"
-	"github.com/jackc/pgx/v5"
 )
 
 func (s *PostgresHoldingStore) GetHoldings(ctx context.Context, userId int, options *types.GetHoldingsStoreOptions) ([]types.Holding, types.PaginationMetadata, error) {
@@ -37,7 +36,7 @@ func (s *PostgresHoldingStore) GetHoldings(ctx context.Context, userId int, opti
 	}
 
 	pgxb := querybuilder.NewPgxQueryBuilder()
-	pgxb.AddQuery("SELECT *, COUNT(*) OVER() as total_items")
+	pgxb.AddQuery(fmt.Sprintf("SELECT %s, COUNT(*) OVER() as total_items", holdingsColumns))
 	pgxb.AddQuery("FROM holdings")
 	pgxb.AddQueryWithPositionals("WHERE user_id = $x", []any{userId})
 
@@ -96,34 +95,4 @@ func (s *PostgresHoldingStore) GetHoldings(ctx context.Context, userId int, opti
 		Page_size:    pageSize,
 		Total_items:  total_items,
 	}, nil
-}
-
-func (s *PostgresHoldingStore) parseRowsIntoHoldings(rows pgx.Rows) ([]types.Holding, int, error) {
-	var total_items int
-	var holdings []types.Holding
-
-	for rows.Next() {
-		var h types.Holding
-		err := rows.Scan(
-			&h.Holding_id,
-			&h.Name,
-			&h.Ticker,
-			&h.Asset_category,
-			&h.Expense_ratio_pct,
-			&h.Maturation_date,
-			&h.Interest_rate_pct,
-			&h.Is_deprecated,
-			&h.User_id,
-			&h.Created_at,
-			&h.Updated_at,
-			&total_items,
-		)
-
-		if err != nil {
-			return nil, 0, err
-		}
-		holdings = append(holdings, h)
-	}
-
-	return holdings, total_items, nil
 }
